# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Viuit Open API",
    "contact": {}
  },
  "host": "127.0.0.1:3000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/oauth/check": {
      "post": {
        "summary": "Check Token",
        "tags": [
          "OAuth"
        ],
        "operationId": "CheckToken",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [{
          "name": "authorization",
          "in": "header",
          "required": true,
          "type": "string",
          "description": ""
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/%5B200%5DCheckToken"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Valid Token"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "ViuIt All Rights Reserved"
              },
              "X-Request-ID": {
                "type": "string",
                "default": "57a0fd77-edfe-40a9-a3d9-fcd95499cf02"
              },
              "Content-Length": {
                "type": "string",
                "default": "25"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"19-ODk1qnsJoqONr7LrYNcftnC1bqM\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 01 Mar 2023 08:55:38 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/%5B401%5DCheckToken1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Invalid Token"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "ViuIt All Rights Reserved"
              },
              "X-Request-ID": {
                "type": "string",
                "default": "5d60a483-ceb1-4a54-a6fe-226dc85e7faa"
              },
              "Content-Length": {
                "type": "string",
                "default": "27"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1b-ptDm3ggC53pgcLgMnSH4sPfBUYY\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 01 Mar 2023 08:55:59 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
  },
...
  "security": [],
...
}